Migrate Query.tokens
