+++ packages/services/api/src/modules/token/resolvers.ts
@@ -12,20 +12,6 @@ const TokenNameModel = z.string().min(2).max(50);
 
 export const resolvers: TokenModule.Resolvers = {
   Query: {
-    async tokens(_, { selector }, { injector }) {
-      const translator = injector.get(IdTranslator);
-      const [organization, project, target] = await Promise.all([
-        translator.translateOrganizationId(selector),
-        translator.translateProjectId(selector),
-        translator.translateTargetId(selector),
-      ]);
-
-      return injector.get(TokenManager).getTokens({
-        organization,
-        project,
-        target,
-      });
-    },
     async tokenInfo(_, __, { injector }) {
       try {
         injector.get(AuthManager).ensureApiToken();
